+++ src/view/dialog.rs
@@ -102,9 +102,9 @@ impl View for Dialog {
         printer.print_box(Vec2::new(0,0), printer.size);
 
         if self.title.len() > 0 {
-            let x = (printer.size.x - self.title.len() as u32) / 2;
+            let x = (printer.size.x - self.title.len()) / 2;
             printer.print((x-2,0), "┤ ");
-            printer.print((x+self.title.len() as u32,0), " ├");
+            printer.print((x+self.title.len(),0), " ├");
 
             printer.with_style(color::TITLE_PRIMARY, |p| p.print((x,0), &self.title));
         }
@@ -131,7 +131,7 @@ impl View for Dialog {
 
         if self.title.len() > 0 {
             // If we have a title, we have to fit it too!
-            inner_size.x = max(inner_size.x, self.title.len() as u32 + 6);
+            inner_size.x = max(inner_size.x, self.title.len() + 6);
         }
 
         inner_size
